11_quiz_app/backend/requirements.txt
# FastAPI Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Snowflake Connector
snowflake-connector-python==3.6.0

# TOML Configuration File Parser
tomli==2.0.1

# CORS Middleware
python-multipart==0.0.6

# Cryptography for Private Key Authentication
cryptography==41.0.7
